refactor: AuthService 디렉터리 변경 반영
